+++ CloudShell.ResourceDefinitions.ReferenceProviders/SqlServer/SqlServerResourceTypeProvider.cs
@@ -21,6 +21,9 @@ public sealed class SqlServerResourceTypeProvider :
 
     public static class Operations
     {
+        public static readonly ResourceOperationId Start = "start";
+        public static readonly ResourceOperationId Stop = "stop";
+        public static readonly ResourceOperationId Restart = "restart";
         public static readonly ResourceOperationId ReconcileAccess =
             "application.sql-server.reconcile-access";
     }
@@ -47,6 +50,9 @@ public sealed class SqlServerResourceTypeProvider :
         },
         Operations:
         [
+            new(Operations.Start),
+            new(Operations.Stop),
+            new(Operations.Restart),
             new(Operations.ReconcileAccess)
         ]);